// src/types/stripe_subscription.interface.ts
export interface IStripeSubscription {
  id: string;
  object: string;
  application: any;
  application_fee_percent: any;
  automatic_tax: AutomaticTax;
  billing_cycle_anchor: number;
  billing_thresholds: any;
  cancel_at: any;
  cancel_at_period_end: boolean;
  canceled_at: any;
  cancellation_details: CancellationDetails;
  collection_method: string;
  created: number;
  currency: string;
  current_period_end: number;
  current_period_start: number;
  customer: string;
  days_until_due: any;
  default_payment_method: string;
  default_source: any;
  default_tax_rates: any[];
  description: any;
  discount: any;
  ended_at: any;
  items: Items;
  latest_invoice: string;
  livemode: boolean;
  metadata: Metadata5;
  next_pending_invoice_item_invoice: any;
  on_behalf_of: any;
  pause_collection: any;
  payment_settings: PaymentSettings;
  pending_invoice_item_interval: any;
  pending_setup_intent: any;
  pending_update: any;
  plan: Plan2;
  quantity: number;
  schedule: any;
  start_date: number;
  status: string;
  test_clock: any;
  transfer_data: any;
  trial_end: any;
  trial_settings: TrialSettings;
  trial_start: any;
}

export interface AutomaticTax {
  enabled: boolean;
}

export interface CancellationDetails {
  comment: any;
  feedback: any;
  reason: any;
}

export interface Items {
  object: string;
  data: Daum[];
  has_more: boolean;
  total_count: number;
  url: string;
}

export interface Daum {
  id: string;
  object: string;
  billing_thresholds: any;
  created: number;
  metadata: Metadata;
  plan: Plan;
  price: Price;
  quantity: number;
  subscription: string;
  tax_rates: any[];
}

export interface Metadata {}

export interface Plan {
  id: string;
  object: string;
  active: boolean;
  aggregate_usage: any;
  amount: number;
  amount_decimal: string;
  billing_scheme: string;
  created: number;
  currency: string;
  interval: string;
  interval_count: number;
  livemode: boolean;
  metadata: Metadata2;
  nickname: any;
  product: string;
  tiers_mode: any;
  transform_usage: any;
  trial_period_days: any;
  usage_type: string;
}

export interface Metadata2 {}

export interface Price {
  id: string;
  object: string;
  active: boolean;
  billing_scheme: string;
  created: number;
  currency: string;
  custom_unit_amount: any;
  livemode: boolean;
  lookup_key: any;
  metadata: Metadata3;
  nickname: any;
  product: Product;
  recurring: Recurring;
  tax_behavior: string;
  tiers_mode: any;
  transform_quantity: any;
  type: string;
  unit_amount: number;
  unit_amount_decimal: string;
}

export interface Metadata3 {}

export interface Product {
  id: string;
  object: string;
  active: boolean;
  attributes: any[];
  created: number;
  default_price: any;
  description: any;
  features: any[];
  images: any[];
  livemode: boolean;
  metadata: Metadata4;
  name: string;
  package_dimensions: any;
  shippable: any;
  statement_descriptor: any;
  tax_code: string;
  type: string;
  unit_label: any;
  updated: number;
  url: any;
}

export interface Metadata4 {
  messagesLimit: string;
}

export interface Recurring {
  aggregate_usage: any;
  interval: string;
  interval_count: number;
  trial_period_days: any;
  usage_type: string;
}

export interface Metadata5 {}

export interface PaymentSettings {
  payment_method_options: any;
  payment_method_types: any;
  save_default_payment_method: string;
}

export interface Plan2 {
  id: string;
  object: string;
  active: boolean;
  aggregate_usage: any;
  amount: number;
  amount_decimal: string;
  billing_scheme: string;
  created: number;
  currency: string;
  interval: string;
  interval_count: number;
  livemode: boolean;
  metadata: Metadata6;
  nickname: any;
  product: string;
  tiers_mode: any;
  transform_usage: any;
  trial_period_days: any;
  usage_type: string;
}

export interface Metadata6 {}

export interface TrialSettings {
  end_behavior: EndBehavior;
}

export interface EndBehavior {
  missing_payment_method: string;
}